requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
altair==5.5.0
analytics-python==1.4.post1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==24.2.0
azure-cognitiveservices-vision-customvision==3.1.1
azure-common==1.1.28
azure-core==1.32.0
azure-mgmt-core==1.5.0
backoff==1.10.0
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cryptography==44.0.0
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.115.6
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.44.1
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
importlib_resources==6.4.5
isodate==0.7.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
mdit-py-plugins==0.4.2
mdurl==0.1.2
monotonic==1.6
msrest==0.7.1
multidict==6.1.0
narwhals==1.17.0
numpy==2.0.2
oauthlib==3.2.2
opencv-python==4.10.0.84
orjson==3.10.12
packaging==24.2
pandas==2.2.3
paramiko==3.5.0
pillow==10.4.0
propcache==0.2.1
protobuf==5.29.1
pyarrow==18.1.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
ruff==0.8.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.3
streamlit==1.41.0
tenacity==9.0.0
toml==0.10.2
tomlkit==0.12.0
tornado==6.4.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
uvicorn==0.32.1
websockets==12.0
yarl==1.18.3
zipp==3.21.0
